Move database import/export callbacks to main activity in order to simplify dependencies
